requirements.txt
requests==2.31.0
gspread==6.2.1
google.auth==2.48.0